Resolve Chapter 2 exercises
This commit provides the answers for all exercises in Chapter 2.
The exercise 2.1 adds the stop condition to the eight queen example to
print only the first solution. Meanwhile, the exercise 2.2 implements
a new method based on all possible permutations and evaluates the
performance of this new method and the old one.

Signed-off-by: Hieu Vu <hieutrgvu@gmail.com>