Resolve Chapter 6 exercises
This commit provides the answers for all exercises in Chapter 6.
This chapter introduces  the features of a Lua function including how
using functions, multiple return results, variadic functions, proper
tail calls. It also shows the usages of table.pack and table.unpack
functions, which are useful when dealing with variadic parameters.

Signed-off-by: Hieu Vu <hieutrgvu@gmail.com>